move create custom KPI to left list in find KPIs section
