+++ src/app/shared/shared-facility-forms/kpi-search-form/add-kpi-search/add-kpi-search.component.css
@@ -18,3 +18,7 @@ hr{
         display: none;
     }
 }
+
+.italic{
+    font-style: italic;
+}
