+++ src/app/shared/shared-facility-forms/kpi-search-form/add-kpi-search/add-kpi-search.component.html
@@ -14,7 +14,7 @@
                 [(ngModel)]="kpiSearchStr">
         </div>
     </div>
-    <div class="col-12">
+    <div class="col-12 mb-1">
         <select name="kpiCategorySearch" class="form-select" id="kpiCategorySearch" [(ngModel)]="kpiCategorySearch">
             <option [ngValue]="undefined">All Categories</option>
             <option *ngFor="let category of primaryKPIs" [ngValue]="category">
@@ -26,6 +26,28 @@
 
 <div class="p-0 w-100">
     <hr>
+</div>
+<div class="d-flex flex-column indicator-container">
+    <div class="d-flex w-100 justify-content-between list-item">
+        <div class="d-flex flex-column">
+            <div>
+                <span class="ps-1 italic">Create Custom KPI</span>
+            </div>
+            <div class="badge-display">
+                <span class="badge bg-secondary">
+                    Custom
+                </span>
+            </div>
+        </div>
+        <div>
+            <a class="p-1 click-link" (click)="openAddCustomModal()">
+                <fa-icon class="small" [icon]="faPlus"></fa-icon>
+            </a>
+        </div>
+    </div>
+    <div class="p-0 w-100">
+        <hr>
+    </div>
 </div>
 <div class="d-flex flex-column indicator-container"
     *ngFor="let option of keyPerformanceIndicatorOptions | kpiListFilter: kpiSearchStr: kpiCategorySearch"
@@ -40,9 +62,6 @@
             </div>
         </div>
         <div>
-            <!-- <a class="p-1 click-link">
-                <fa-icon class="small" [icon]="faCircleQuestion"></fa-icon>
-            </a> -->
             <a class="p-1 click-link" (click)="addKPI(option)">
                 <fa-icon class="small" [icon]="faPlus"></fa-icon>
             </a>
@@ -50,5 +69,45 @@
     </div>
     <div class="p-0 w-100">
         <hr>
+    </div>
+</div>
+
+
+
+
+
+<!--custom KPI modal-->
+<div [ngClass]="{'window-overlay': displayCustomKPIModal}"></div>
+<div class="popup" [ngClass]="{'open': displayCustomKPIModal }">
+    <div class="popup-header">Create Custom KPI?
+        <button type="button" class="btn-close float-end" aria-label="Close" (click)="closeCustomKPIModal()"></button>
+    </div>
+    <div class="popup-body">
+        <form>
+            <div class="row mb-1">
+                <label class="col-sm-5 col-form-label" for="customKPIName">Custom KPI Name</label>
+                <div class="col-sm-7">
+                    <input name="customKPIName" type="text" class="form-control" [(ngModel)]="customKPIName"
+                        id="customKPIName">
+                    <div class="alert alert-danger small mt-1 p-1" *ngIf="!customKPIName">
+                        Name Required
+                    </div>
+                </div>
+            </div>
+            <div class="row mb-1">
+                <label class="col-sm-5 col-form-label" for="kpiCategory">KPI Category</label>
+                <div class="col-sm-7">
+                    <select class="form-select" name="kpiCategory" id="kpiCategory" [(ngModel)]="kpiCategory">
+                        <option *ngFor="let primaryKPI of primaryKPIs" [ngValue]="primaryKPI">{{primaryKPI}}</option>
+                    </select>
+                </div>
+            </div>
+        </form>
+        <hr>
+    </div>
+    <div class="popup-footer d-flex justify-content-end">
+        <button class="btn btn-secondary btn-sm me-2" (click)="closeCustomKPIModal()">Cancel</button>
+        <button class="btn btn-success btn-sm" [disabled]="!customKPIName" (click)="confirmCreate()">Create Custom
+            KPI</button>
     </div>
 </div>
