+++ src/app/shared/shared-facility-forms/kpi-search-form/kpi-list/kpi-list.component.css
@@ -1,9 +1,3 @@
-.list-item {
-    /* border: solid 1px;
-    padding: 2px; */
-}
-
-
 .click-link-danger {
     color: red;
 }
