+++ src/app/shared/shared-facility-forms/kpi-search-form/kpi-list/kpi-list.component.ts
@@ -1,19 +1,16 @@
-import { Component, Input } from '@angular/core';
+import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 import { IconDefinition, faBullseye, faCircleQuestion, faPlus, faTrash, faUser } from '@fortawesome/free-solid-svg-icons';
-import { Subscription, firstValueFrom } from 'rxjs';
-import { CompanyIdbService } from 'src/app/indexed-db/company-idb.service';
+import { Subscription } from 'rxjs';
 import { DbChangesService } from 'src/app/indexed-db/db-changes.service';
 import { FacilityIdbService } from 'src/app/indexed-db/facility-idb.service';
 import { KeyPerformanceIndicatorsIdbService } from 'src/app/indexed-db/key-performance-indicators-idb.service';
 import { KeyPerformanceMetricImpactsIdbService } from 'src/app/indexed-db/key-performance-metric-impacts-idb.service';
 import { OnSiteVisitIdbService } from 'src/app/indexed-db/on-site-visit-idb.service';
-import { IdbCompany } from 'src/app/models/company';
 import { IdbFacility } from 'src/app/models/facility';
-import { IdbKeyPerformanceIndicator, getNewKeyPerformanceIndicator } from 'src/app/models/keyPerformanceIndicator';
+import { IdbKeyPerformanceIndicator } from 'src/app/models/keyPerformanceIndicator';
 import { IdbKeyPerformanceMetricImpact } from 'src/app/models/keyPerformanceMetricImpact';
 import { IdbOnSiteVisit } from 'src/app/models/onSiteVisit';
-import { KeyPerformanceIndicatorOption, PrimaryKPI, PrimaryKPIs } from 'src/app/shared/constants/keyPerformanceIndicatorOptions';
 
 @Component({
     selector: 'app-kpi-list',
@@ -39,15 +36,9 @@ export class KpiListComponent {
   keyPerformanceIndicators: Array<IdbKeyPerformanceIndicator>;
   keyPerformanceIndicatorSub: Subscription;
 
-  displayCustomKPIModal: boolean = false;
-  customKPIName: string = '';
-  primaryKPIs: Array<PrimaryKPI> = PrimaryKPIs;
-  kpiCategory: PrimaryKPI = 'Other';
-
   keyPerformanceMetricImpacts: Array<IdbKeyPerformanceMetricImpact>
   keyPerformanceMetricImpactsSub: Subscription;
   constructor(
-    private companyIdbService: CompanyIdbService,
     private keyPerformanceIndicatorIdbService: KeyPerformanceIndicatorsIdbService,
     private dbChangesService: DbChangesService,
     private router: Router,
@@ -88,27 +79,6 @@ export class KpiListComponent {
   async removeKPI() {
     await this.dbChangesService.deleteKPIs([this.kpiToDelete]);
     this.closeDeleteModal();
-  }
-
-  openAddCustomModal() {
-    this.displayCustomKPIModal = true;
-  }
-
-  closeCustomKPIModal() {
-    this.displayCustomKPIModal = false;
-  }
-
-  async confirmCreate() {
-    let option: KeyPerformanceIndicatorOption = {
-      primaryKPI: this.kpiCategory,
-      label: this.customKPIName,
-      htmlLabel: this.customKPIName,
-      optionValue: 'other',
-    };
-    let newKPI: IdbKeyPerformanceIndicator = getNewKeyPerformanceIndicator(this.facility.userId, this.facility.companyId, option, true, this.facility.guid);
-    await firstValueFrom(this.keyPerformanceIndicatorIdbService.addWithObservable(newKPI));
-    await this.keyPerformanceIndicatorIdbService.setKeyPerformanceIndicators();
-    this.closeCustomKPIModal();
   }
 
   goToKpiDetails(kpi: IdbKeyPerformanceIndicator) {
